refactor: localize exception handling
